+++ bashly/commands/upgrade.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+WORKSPACE_YES="$(workspace_yes)"
+upgrade_workspace "$WORKSPACE_YES"